+++ CHANGELOG.md
@@ -12,6 +12,7 @@ v0.5.0 (in development)
     - Added `--package` option for selecting the package to update
     - Added `--workspace` option for setting MSRV workspace-wide
 - Added workspace support to `mkgithub`
+- `mkgithub`: Added `--plan-only` option
 
 v0.4.0 (2024-12-17)
 -------------------